# Conflicts:
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/fragment/DynamicProductDetailFragment.kt

+++ features/user/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/viewmodel/SendableProductPreview.kt
@@ -37,7 +37,7 @@ class SendableProductPreview(
 
     private fun generateResultProduct(): ResultProduct {
         return ResultProduct(
-                productPreview.id.toInt(),
+                productPreview.id,
                 productPreview.url,
                 productPreview.imageUrl,
                 productPreview.price,
